Splits shopping_spec
